requirements.txt
```txt
streamlit
opencv-python-headless
opencv-contrib-python-headless
numpy
pillow
imutils